software/robot/software/TWIPR-Software/requirements.txt
pygame~=2.6.0
evdev~=1.7.1
pillow~=11.0.0
graphviz~=0.20.3
cobs~=1.2.1
orjson~=3.10.5
pyserial~=3.5
numpy~=2.0.0
stm32loader~=0.7.1
gTTs ~= 2.5.4
luma.oled ~= 3.14.0
edge-tts ~= 7.0.0
pydub ~= 0.25.1
pyttsx3 ~= 2.98